requirements.txt
attrs==20.3.0
data==0.4
decorator==5.0.7
funcsigs==1.0.2
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
